Changed sending and CRC function
Everybody is using the other CRC function, so they made us switch. I've
changed the packetize functions to use this.

We were thinking of using a thread to send a file, and wait until the
buffer has packets in it and then send. This isnt working so whenever we
open a file the send thread will be created if it isnt already running
and send the packets.